Skip the "Website URL" message at the end of deploy on failing getPullZone.
When getPullZone caused a 401 (Unauthorized) response, json_decode would produce null (as there is no response body), and thus $result->getData()->Hostnames would fail. An if-block was added to prevent this, at the expense of the "Website URL" message. This may be related to https://github.com/own3d/bunny-action/issues/2.